Moved method to a separate method and added tests
